requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
numpy-financial>=1.0.0
plotly>=5.18.0
openpyxl>=3.1.0
google-generativeai>=0.5.0
reportlab>=4.0.0
requests>=2.31.0
gspread>=6.0.0
oauth2client>=4.1.3
